Abstract
Implementation of QPSK transmitter and receiver using NI USRP

ABSTRACT:
We are going to implement a system with Quadrature phase shift keying (QPSK) transmitter and receiver using NI USRP and LabVIEW software. QPSK is one of the most used Phase shift keying modulation technique in wireless communication because of its ability to transmit twice the data rate for a given bandwidth. PSK is considered to be very resistant to nonlinear effects since the envelope of the signal is constant. Here we are going to use LABVIEW (Laboratory Virtual Instrumentation Engineering Workbench) as the simulation platform because the programming environment being graphical, gives the good visualization of the results. And in final setup we receive a signal in presence of AWGN with the help of designed transmitter and receiver.

TIMELINE: 
OCT 1st week: Intro to LabVIEW (since LabVIEW is new for us so, to get familiar with its tool and environment we invest our first week in exploring the dimensions of LabVIEW). 
OCT 2nd week: Implementation of QPSK modulator for transmitter side in LABVIEW. 
OCT 3rd week: Implementation of QPSK modulator for transmitter side in LABVIEW.. 
OCT 4th week: Testing of setup by adding AWGN in a signal.